% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_searches.R
\name{check_coords_for_error}
\alias{check_coords_for_error}
\title{Validates location coordinates}
\usage{
check_coords_for_error(coords)
}
\arguments{
\item{coords}{Location coordinates. Must use this format: list(lat = 0, lng = 0)}
}
\value{
TRUE if coords are valid, FALSE otherwise
}
\description{
Validates location coordinates
}
