% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding_reverse.R
\name{geocoding_reverse}
\alias{geocoding_reverse}
\title{Reverse Geocoding}
\usage{
geocoding_reverse(lat, lng)
}
\arguments{
\item{lat}{Latitude of the point to reverse geocode.}

\item{lng}{Longitude of the point to reverse geocode.}
}
\value{
API response parsed as list and as a raw json
}
\description{
Attempt to match a latitude, longitude pair to an address.
}
\details{
See \url{https://traveltime.com/docs/api/reference/geocoding-reverse} for details
}
\examples{
\dontrun{
geocoding_reverse(lat=51.507281, lng=-0.132120)
}
}
