% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geohash_fast_proto.R
\name{geohash_fast_proto}
\alias{geohash_fast_proto}
\title{Geohash (Fast) with Protobuf}
\usage{
geohash_fast_proto(
  lat,
  lng,
  country = c(
    "nl", "at", "uk", "be", "de", "fr", "ie", "lt", "us", "za",
    "ro", "pt", "ph", "nz", "no", "lv", "jp", "in", "id", "hu",
    "gr", "fi", "dk", "ca", "au", "sg", "ch", "es", "it", "pl",
    "se", "li", "mx", "sa", "rs", "si", "cz", "sk"
  ),
  travelTime,
  transportation = names(protoTransport),
  resolution,
  type = c("one_to_many", "many_to_one"),
  properties = NULL
)
}
\arguments{
\item{lat}{origin/destination latitude}

\item{lng}{origin/destination longitude}

\item{country}{Origin country. See \url{https://docs.traveltime.com/api/overview/supported-countries} for the list of supported countries.}

\item{travelTime}{Maximum journey time (in seconds).}

\item{transportation}{One of supported transportation methods: 'pt', 'driving+ferry', 'cycling+ferry', 'walking+ferry'.}

\item{resolution}{Geohash resolution level.}

\item{type}{Request type: 'one_to_many' or 'many_to_one'.}

\item{properties}{Optional list of cell property types to include in the response (e.g. 0 = MIN, 1 = MAX, 2 = MEAN).}
}
\value{
API response parsed as a list and as a raw json
}
\description{
Returns geohash cells that are reachable from the origin within the given travel time.
Uses Protocol Buffers for higher performance.
}
\examples{
\dontrun{
geohash_fast_proto(
  lat = 51.508930,
  lng = -0.131387,
  transportation = 'driving+ferry',
  travelTime = 1800,
  country = "uk",
  resolution = 6
)
}
}
