% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_searches.R
\name{make_location}
\alias{make_location}
\title{Location objects constructor}
\usage{
make_location(id, coords)
}
\arguments{
\item{id}{You will have to reference this id in your searches.
It will also be used in the response body. MUST be unique among all locations.}

\item{coords}{Location coordinates. Must use this format: list(lat = 0, lng = 0)}
}
\value{
A data.frame wrapped in a list.
It is constructed in a way that allows jsonlite::toJSON to correctly transform it into a valid request body
}
\description{
Define your locations to use later in departure_searches or arrival_searches.
}
\details{
See \url{https://docs.traveltime.com/api/reference/distance-matrix} for details
}
\seealso{
See \code{\link{time_filter}} for usage examples
}
