% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_searches.R
\name{make_union_intersect}
\alias{make_union_intersect}
\title{Set objects constructor}
\usage{
make_union_intersect(id, search_ids)
}
\arguments{
\item{id}{Used to identify this specific search in the results array. MUST be unique among all searches.}

\item{search_ids}{An unnamed list of search ids which results will formulate this union.}
}
\value{
A data.frame wrapped in a list.
It is constructed in a way that allows jsonlite::toJSON to correctly transform it into a valid request body
}
\description{
Allows you to define unions or intersections of shapes that are results of previously defined searches.
You can define a maximum of 10 unions/intersections
}
\details{
See \url{https://docs.traveltime.com/api/reference/isochrones} for details
}
\seealso{
See \code{\link{time_map}} for usage examples
}
