% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-info.R
\name{map_info}
\alias{map_info}
\title{Map Info}
\usage{
map_info()
}
\value{
API response parsed as list and as a raw json
}
\description{
Returns information about currently supported countries.
}
\details{
See \url{https://docs.traveltime.com/api/reference/map-info/} for details
}
\examples{
\dontrun{
map_info()
}
}
