% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{geocoding}
\alias{geocoding}
\title{Geocoding (Search)}
\usage{
geocoding(query, within.country = NA, exclude.location.types = NA)
}
\arguments{
\item{query}{A query to geocode. Can be an address, a postcode or a venue.}

\item{within.country}{Only return the results that are within the specified country.
If no results are found it will return the country itself. Optional. Format:ISO 3166-1 alpha-2 or alpha-3}

\item{exclude.location.types}{Exclude location types from results. Available values: "country". Optional.}
}
\value{
API response parsed as list and as a raw json
}
\description{
Match a query string to geographic coordinates.
}
\details{
See \url{https://traveltime.com/docs/api/reference/geocoding-search} for details
}
\examples{
geocoding('Parliament square')
}
