% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding_reverse.R
\name{geocoding_reverse}
\alias{geocoding_reverse}
\title{Reverse Geocoding}
\usage{
geocoding_reverse(lat, lng, within.country = NA, exclude.location.types = NA)
}
\arguments{
\item{lat}{Latitude of the point to reverse geocode.}

\item{lng}{Longitude of the point to reverse geocode.}

\item{within.country}{Only return the results that are within the specified country.
If no results are found it will return the country itself. Optional. Format:ISO 3166-1 alpha-2 or alpha-3}
}
\value{
API response parsed as list and as a raw json
}
\description{
Attempt to match a latitude, longitude pair to an address.
}
\details{
See \url{https://traveltime.com/docs/api/reference/geocoding-reverse} for details
}
\examples{
\dontrun{
geocoding_reverse(lat=51.507281, lng=-0.132120)
}
}
